// fishbangla-web/src/app/interfaces/fishproduct.ts
export interface FishProduct {
    productId: string;
    upazillaName: string;
    districtName: string;
    userFullName: string;
    unitSize: string;
    divisionName: string;
    appMenuTitle: string;
    maxUnitQty: string;
    userId: string;
    productName: string;
    avgWeightUnit: string;
    unitType: string;
    productToken: string;
    productPostType: string;
    price: string;
    productPicture: string;
    productVideo: string;
    avgWeight: string;
    minUnitQty: string;
    productRating: string;
    isFavourite: string;
}